+++ apps/backoffice-v2/src/lib/blocks/create-blocks-typed/types.ts
@@ -41,7 +41,7 @@ export type TContainerCell = {
 
 export type THeadingCell = {
   type: 'heading';
-  value: string;
+  value: ReactNode;
   props?: ComponentProps<'h2'>;
 };
 
@@ -239,10 +239,18 @@ export type TNodeCell = {
 
 export type TReadOnlyDetailsCell = {
   type: 'readOnlyDetails';
-  props?: ComponentProps<'div'> & Pick<ComponentProps<typeof ReadOnlyDetail>, 'parse'>;
+  props?: ComponentProps<'div'> & {
+    config?: Pick<ComponentProps<typeof ReadOnlyDetail>, 'parse'> & {
+      sort?: {
+        direction?: SortDirection;
+        predefinedOrder?: string[];
+      };
+    };
+  };
   value: Array<{
     label: string;
     value: ExtendedJson;
+    props?: Omit<ComponentProps<typeof ReadOnlyDetail>, 'children'>;
   }>;
 };