+++ packages/ui/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @ballerine/ui
 
+## 0.7.136
+
+### Patch Changes
+
+- Bump
+
+## 0.7.135
+
+### Patch Changes
+
+- version bump
+- Updated dependencies
+  - @ballerine/common@0.9.91
+
+## 0.7.134
+
+### Patch Changes
+
+- bump
+- Updated dependencies
+  - @ballerine/common@0.9.90
+
 ## 0.7.133
 
 ### Patch Changes